Change defaults (private) method.
Remove line in defaults (private) method and put the same
functionality in the method that it calls, config (private). Also,
get rid of comma in the defaults method's return hash. Remove a
comma after the variable 'extension' in that return hash, as it
causes confusion as to whether that var. is referencing the first
element of an array instead of a string.