Fix the gist.copy config option
str_to_bool already handles string values such as 'true' and 'on'.
Trying to do a regexp match on the boolean variable returned from the
config method will just result in that value to always be false, which
means that the gist.copy config value doesn't work as expected.
Removing this check fixes the problem.