Chomp OAuth access token before use.
If a user is modifying the .gist file by hand there is a big risk that the
user adds a newline to the file. That newline is causing gist to fail with
"Error: Got Net::HTTPUnauthorized from gist:".
So better remove the eventual newline with chomp before use.

This is extra handy when setting up gist on an computer that you don't trust
with your real GitHub credentials (pre-generated OAuth token).