Load GitHub token from netrc if it's not in AuthTokenFile
Git, curl, and other tools already offer support for loading credentials
from ~/.netrc, namely when interacting with GitHub. This uses the netrc
gem to play along. It's implemented as a fallback, in case a token is
not found in the gist-specific file.

To avoid adding gem dependencies, the netrc library is imported from
https://github.com/heroku/netrc into the vendor folder. It is licensed
under the MIT license, the same as this project, and it's copyright
notice and license have been embedded in a comment with the code.

The standalone script build/gist was regenerated and pulls in missing
changes in addition to the netrc dependency.

Fixes #260.